+++ src/y2019/day7/Part1.java
@@ -3,27 +3,27 @@ package y2019.day7;
 import utils.Problem;
 import y2019.intcode.Parser;
 
-public class Part1 extends Problem<Long> {
+public class Part1 extends Problem<Integer> {
     private Parser parser;
 
-    public Part1(String fileName, long exp) {
+    public Part1(String fileName, int exp) {
         super(exp);
 
         this.parser = new Parser(fileName);
     }
 
-    public Long run() {
+    public Integer run() {
         var prog = parser.parse();
         var perms = Utils.findPerms(new int[] { 0, 1, 2, 3, 4 });
 
-        var most = 0L;
+        var most = 0;
 
         for (var perm : perms) {
             var circuit = new Circuit(prog, 5);
 
             circuit.init(perm);
 
-            var signal = circuit.run(0L);
+            var signal = circuit.run(0);
             if (signal > most) {
                 most = signal;
             }
